refactor(app): extract helpers and add Opal # await: on app entry
- Add app/helpers/{session_cookie,chat_history,markdown_render}.rb (PLAN B1)
- Wire JwtKeyHelpers, binding/chat helpers, markdown_html; include session cookie instance methods
- Keep App-level constants (JWT/CHAT/session) for route constant lookup
- Regenerate fragments to use markdown_html helper; update canonical_all.rb accordingly

Made-with: Cursor
