+++ app/helpers/markdown_render.rb
@@ -1,0 +1,10 @@
+# await: all, authenticate!, call, chat_verify_token!, clear_chat_history, decode, dh_compute_key, dispatch_js, dispatch_scheduled, encode, execute, execute_insert, fetch, fetch_raw, final, get_binary, get_first_row, get_response, list, load_chat_history, open, private_decrypt, public_encrypt, run, save_chat_history, send, sign, sign_pss, sleep, verify, verify_pss
+# frozen_string_literal: true
+
+module Homurabi
+  module MarkdownRenderHelpers
+    def markdown_html(text)
+      HomurabiMarkdown.render(text)
+    end
+  end
+end
