+++ app/routes/canonical_all.rb
@@ -909,7 +909,7 @@
       # lists, bold, code fences, links). Safe to insert because
       # `HomurabiMarkdown.render` HTML-escapes the input first and
       # restricts link hrefs to http/https/mailto/relative.
-      'reply_html'   => HomurabiMarkdown.render(reply_text),
+      'reply_html'   => markdown_html(reply_text),
       'history_len'  => new_history.size
     }.to_json
   end
@@ -935,7 +935,7 @@
       item = { 'role' => role, 'content' => content }
       # Only assistant replies are converted — user messages are
       # authored text and stay as-is to preserve the exact payload.
-      item['content_html'] = HomurabiMarkdown.render(content) if role == 'assistant'
+      item['content_html'] = markdown_html(content) if role == 'assistant'
       item
     end
     {
